CURLOPT_TIMEOUT is a long!
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10328 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/common/URLManager.cxx
@@ -121,7 +121,7 @@ bool URLManager::getURLHeader(const std::string& URL)
   }
 #endif
 
-  result = curl_easy_setopt(curl, CURLOPT_TIMEOUT, timeout);
+  result = curl_easy_setopt(curl, CURLOPT_TIMEOUT, (long)timeout);
   if (result != CURLE_OK) {
     errorCode = result;
     DEBUG1("CURLOPT_TIMEOUT error: %d\n", result);
